requirements.txt
flask==2.2.2
flask-cors==3.0.10
flask-login==0.6.3
flask-wtf==1.1.1
flask-limiter==3.5.0
werkzeug==2.2.3
python-pptx==0.6.21
anthropic==0.25.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pdfplumber==0.9.0
reportlab==4.0.4
pdf2image==1.16.3
Pillow==10.0.0
tenacity==8.2.0
PyMuPDF==1.23.8
matplotlib==3.7.2
bcrypt==4.0.1
email-validator==2.0.0